Local Account creation
Modified the KRB plugin to create a local account for the user, and the
configuration window for this plugin to allow for various local account
options.  This bypasses a bug in the pGina system that is failing logins
on the first attempt because there is no local account getting created
correctly.